Move utils to vf utils
